[BH-1452] Remove board directory from module-os
Moved board-specific code to module-bsp.
Extracted several smaller CMake targets to avoid
circural dependencies and reduce compile time.
