+++ module-bsp/board/linux/os/FreeRTOSConfig.h
@@ -75,7 +75,7 @@ extern "C"
 /* Memory allocation related definitions. */
 #define configSUPPORT_STATIC_ALLOCATION  1
 #define configSUPPORT_DYNAMIC_ALLOCATION 1
-#define configTOTAL_HEAP_SIZE ((size_t)(1024 * 512))
+#define configTOTAL_HEAP_SIZE            ((size_t)(1024 * 512))
 
 #define configAPPLICATION_ALLOCATED_HEAP 0
