--- module-os/board/rt1051/atexit.h
@@ -1,15 +1,0 @@
-/*
- *	Common definitions for atexit-like routines
- */
-
-enum __atexit_types
-{
-  __et_atexit,
-  __et_onexit,
-  __et_cxa
-};
-
-#define MALLOC_PROVIDED
-
-void __call_exitprocs (int, void *);
-int __register_exitproc (int, void (*fn) (void), void *, void *);
